+++ assembly/src/ast/parsers/adv_ops.rs
@@ -33,6 +33,10 @@ pub fn parse_adv_inject(op: &Token) -> Result<Node, ParsingError> {
             2 => AdvInject(PushSmtGet),
             _ => return Err(ParsingError::extra_param(op)),
         },
+        "push_smtinsert" => match op.num_parts() {
+            2 => AdvInject(PushSmtInsert),
+            _ => return Err(ParsingError::extra_param(op)),
+        },
         "push_mapval" => match op.num_parts() {
             2 => AdvInject(PushMapVal),
             3 => {
@@ -75,6 +79,10 @@ pub fn parse_adv_inject(op: &Token) -> Result<Node, ParsingError> {
                     AdvInject(InsertHdwordImm { domain })
                 }
             }
+            _ => return Err(ParsingError::extra_param(op)),
+        },
+        "insert_hperm" => match op.num_parts() {
+            2 => AdvInject(InsertHperm),
             _ => return Err(ParsingError::extra_param(op)),
         },
         _ => return Err(ParsingError::invalid_op(op)),