+++ processor/src/decorators/adv_stack_injectors.rs
@@ -1,7 +1,8 @@
 use super::{super::Ext2InttError, AdviceProvider, AdviceSource, ExecutionError, Process};
 use vm_core::{
-    crypto::merkle::EmptySubtreeRoots, utils::collections::Vec, Felt, FieldElement, QuadExtension,
-    StarkField, Word, ONE, ZERO,
+    crypto::{hash::RpoDigest, merkle::EmptySubtreeRoots},
+    utils::collections::Vec,
+    Felt, FieldElement, QuadExtension, StarkField, Word, ONE, WORD_SIZE, ZERO,
 };
 use winter_prover::math::fft;
 
@@ -338,6 +339,225 @@ where
 
         Ok(())
     }
+
+    /// Pushes values onto the advice stack which are required for successful insertion of a
+    /// key-value pair into a Sparse Merkle Tree data structure.
+    ///
+    /// The Sparse Merkle Tree is tiered, meaning it will have leaf depths in `{16, 32, 48, 64}`.
+    ///
+    /// Inputs:
+    ///   Operand stack: [VALUE, KEY, ROOT, ...]
+    ///   Advice stack: [...]
+    ///
+    /// Outputs:
+    ///   Operand stack: [OLD_VALUE, NEW_ROOT, ...]
+    ///   Advice stack: see comments for specialized handlers below.
+    ///
+    /// Where:
+    /// - ROOT and NEW_ROOT are the roots of the TSMT before and after the insert respectively.
+    /// - VALUE is the value to be inserted.
+    /// - OLD_VALUE is the value previously associated with the specified KEY.
+    ///
+    /// # Errors
+    /// Will return an error if:
+    /// - The Merkle store does not contain a node with the specified root.
+    /// - The Merkle store does not contain all nodes needed to validate the path between the root
+    ///   and the relevant TSMT nodes.
+    /// - The advice map does not contain required data about TSMT leaves to be modified.
+    ///
+    /// # Panics
+    /// Will panic as unimplemented if the target depth is `64`.
+    pub(super) fn push_smtinsert_inputs(&mut self) -> Result<(), ExecutionError> {
+        // get the key and tree root from the stack
+        let key = self.stack.get_word(1);
+        let root = self.stack.get_word(2);
+
+        // determine the depth of the first leaf or an empty tree node
+        let index = &key[3];
+        let depth = self.advice_provider.get_leaf_depth(root, &SMT_MAX_TREE_DEPTH, index)?;
+        debug_assert!(depth < 65);
+
+        // map the depth value to its tier; this rounds up depth to 16, 32, 48, or 64
+        let depth = SMT_NORMALIZED_DEPTHS[depth as usize];
+        if depth == 64 {
+            unimplemented!("handling of depth=64 tier hasn't been implemented yet");
+        }
+
+        // get the value of the node a this index/depth
+        let index = index.as_int() >> (64 - depth);
+        let index = Felt::new(index);
+        let node = self.advice_provider.get_tree_node(root, &Felt::new(depth as u64), &index)?;
+
+        // figure out what kind of insert we are doing; possible options are:
+        // - if the node is a root of an empty subtree, this is a simple insert.
+        // - if the node is a leaf, this could be either an update (for the same key), or a
+        //   complex insert (i.e., the existing leaf needs to be moved to a lower tier).
+        let empty = EmptySubtreeRoots::empty_hashes(64)[depth as usize];
+        if node == Word::from(empty) {
+            self.handle_smt_simple_insert(root, depth, index)?;
+        } else {
+            // get the key and value stored in the current leaf
+            let (leaf_key, leaf_value) = self.get_smt_upper_leaf_preimage(node)?;
+
+            // if the key for the value to be inserted is the same as the leaf's key, we are
+            // dealing with a simple update; otherwise, we are dealing with a complex insert
+            if leaf_key == key {
+                self.handle_smt_update(depth, leaf_value)?;
+            } else {
+                self.handle_smt_complex_insert(depth, key, leaf_key, leaf_value)?;
+            }
+        }
+
+        Ok(())
+    }
+
+    // TSMT INSERT HELPER METHODS
+    // --------------------------------------------------------------------------------------------
+
+    /// Retrieves a key-value pair for the specified leaf node from the advice map.
+    ///
+    /// # Errors
+    /// Returns an error if the value under the specified node does not exist or does not consist
+    /// of exactly 8 elements.
+    fn get_smt_upper_leaf_preimage(&self, node: Word) -> Result<(Word, Word), ExecutionError> {
+        let node_bytes = RpoDigest::from(node).as_bytes();
+        let kv = self
+            .advice_provider
+            .get_mapped_values(&node_bytes)
+            .ok_or(ExecutionError::AdviceMapKeyNotFound(node))?;
+
+        if kv.len() != WORD_SIZE * 2 {
+            return Err(ExecutionError::AdviceMapValueInvalidLength(node, WORD_SIZE * 2, kv.len()));
+        }
+
+        let key = [kv[0], kv[1], kv[2], kv[3]];
+        let val = [kv[4], kv[5], kv[6], kv[7]];
+        Ok((key, val))
+    }
+
+    /// Prepares the advice stack for a TSMT update operation. Specifically, the advice stack will
+    /// be arranged as follows:
+    ///
+    /// - [ZERO (padding), d0, d1, ONE (is_update), OLD_VALUE]
+    ///
+    /// Where:
+    /// - d0 is a boolean flag set to `1` if the depth is `16` or `48`.
+    /// - d1 is a boolean flag set to `1` if the depth is `16` or `32`.
+    /// - OLD_VALUE is the current value in the leaf to be updated.
+    fn handle_smt_update(&mut self, depth: u8, old_value: Word) -> Result<(), ExecutionError> {
+        // put the old value onto the advice stack
+        self.advice_provider.push_stack(AdviceSource::Word(old_value))?;
+
+        // set is_update flag to ONE
+        self.advice_provider.push_stack(AdviceSource::Value(ONE))?;
+
+        // set depth flags based on leaf's depth
+        let (is_16_or_32, is_16_or_48) = get_depth_flags(depth);
+        self.advice_provider.push_stack(AdviceSource::Value(is_16_or_32))?;
+        self.advice_provider.push_stack(AdviceSource::Value(is_16_or_48))?;
+
+        // pad the advice stack with an extra value to make it consistent with other cases when
+        // we expect 4 flag values on the top of the advice stack
+        self.advice_provider.push_stack(AdviceSource::Value(ZERO))?;
+
+        Ok(())
+    }
+
+    /// Prepares the advice stack for a TSMT simple insert operation (i.e., when we are replacing
+    /// an empty node). Specifically, the advice stack will be arranged as follows:
+    ///
+    /// - Simple insert at depth 16: [d0, d1, ONE (is_simple_insert), ZERO (is_update)]
+    /// - Simple insert at depth 32 or 48: [d0, d1, ONE (is_simple_insert), ZERO (is_update), P_NODE]
+    ///
+    /// Where:
+    /// - d0 is a boolean flag set to `1` if the depth is `16` or `48`.
+    /// - d1 is a boolean flag set to `1` if the depth is `16` or `32`.
+    /// - P_NODE is an internal node located at the tier above the insert tier.
+    fn handle_smt_simple_insert(
+        &mut self,
+        root: Word,
+        depth: u8,
+        index: Felt,
+    ) -> Result<(), ExecutionError> {
+        // put additional data onto the advice stack as needed
+        match depth {
+            16 => (), // nothing to do; all the required data is already in the VM
+            32 | 48 => {
+                // for depth 32 and 48, we need to provide the internal node located on the tier
+                // above the insert tier
+                let p_index = Felt::from(index.as_int() >> 16);
+                let p_depth = Felt::from(depth - 16);
+                let p_node = self.advice_provider.get_tree_node(root, &p_depth, &p_index)?;
+                self.advice_provider.push_stack(AdviceSource::Word(p_node))?;
+            }
+            64 => unimplemented!("insertions at depth 64 are not yet implemented"),
+            _ => unreachable!("invalid depth {depth}"),
+        }
+
+        // push is_update and is_simple_insert flags onto the advice stack
+        self.advice_provider.push_stack(AdviceSource::Value(ZERO))?;
+        self.advice_provider.push_stack(AdviceSource::Value(ONE))?;
+
+        // set depth flags based on node's depth
+        let (is_16_or_32, is_16_or_48) = get_depth_flags(depth);
+        self.advice_provider.push_stack(AdviceSource::Value(is_16_or_32))?;
+        self.advice_provider.push_stack(AdviceSource::Value(is_16_or_48))?;
+
+        Ok(())
+    }
+
+    /// Prepares the advice stack for a TSMT complex insert operation (i.e., when a leaf node needs
+    /// to be replaced with a subtree of nodes at a lower tier). Specifically, the advice stack
+    /// will be arranged as follows:
+    ///
+    ///  - [d0, d1, ZERO (is_simple_insert), ZERO (is_update), E_KEY, E_VALUE]
+    ///
+    /// Where:
+    /// - d0 and d1 are boolean flags a combination of which determines the source and the target
+    ///   tiers as follows:
+    ///   - (0, 0): depth 16 -> 32
+    ///   - (0, 1): depth 16 -> 48
+    ///   - (1, 0): depth 32 -> 48
+    ///   - (1, 1): depth 16, 32, or 48 -> 64
+    /// - E_KEY and E_VALUE are the key-value pair for a leaf which is to be replaced by a subtree.
+    fn handle_smt_complex_insert(
+        &mut self,
+        depth: u8,
+        key: Word,
+        leaf_key: Word,
+        leaf_value: Word,
+    ) -> Result<(), ExecutionError> {
+        // push the key and value onto the advice stack
+        self.advice_provider.push_stack(AdviceSource::Word(leaf_value))?;
+        self.advice_provider.push_stack(AdviceSource::Word(leaf_key))?;
+
+        // push is_update and is_simple_insert flags onto the advice stack
+        self.advice_provider.push_stack(AdviceSource::Value(ZERO))?;
+        self.advice_provider.push_stack(AdviceSource::Value(ZERO))?;
+
+        // determine the combination of the source and target tiers for the insert
+        // and populate the depth flags accordingly
+        let common_prefix = get_common_prefix(&key, &leaf_key);
+        let target_depth = SMT_NORMALIZED_DEPTHS[common_prefix as usize + 1];
+        match target_depth {
+            32 if depth == 16 => {
+                self.advice_provider.push_stack(AdviceSource::Value(ONE))?;
+                self.advice_provider.push_stack(AdviceSource::Value(ONE))?;
+            }
+            48 if depth == 16 => {
+                self.advice_provider.push_stack(AdviceSource::Value(ONE))?;
+                self.advice_provider.push_stack(AdviceSource::Value(ZERO))?;
+            }
+            48 if depth == 32 => {
+                self.advice_provider.push_stack(AdviceSource::Value(ZERO))?;
+                self.advice_provider.push_stack(AdviceSource::Value(ONE))?;
+            }
+            64 => unimplemented!("insertions at depth 64 are not yet implemented"),
+            _ => unreachable!("invalid source/target tier combination: {depth} -> {target_depth}"),
+        }
+
+        Ok(())
+    }
 }
 
 // HELPER FUNCTIONS
@@ -347,4 +567,16 @@ fn u64_to_u32_elements(value: u64) -> (Felt, Felt) {
     let hi = Felt::new(value >> 32);
     let lo = Felt::new((value as u32) as u64);
     (hi, lo)
+}
+
+fn get_common_prefix(key1: &Word, key2: &Word) -> u8 {
+    let k1 = key1[3].as_int();
+    let k2 = key2[3].as_int();
+    (k1 ^ k2).leading_zeros() as u8
+}
+
+fn get_depth_flags(depth: u8) -> (Felt, Felt) {
+    let is_16_or_32 = if depth == 16 || depth == 32 { ONE } else { ZERO };
+    let is_16_or_48 = if depth == 16 || depth == 48 { ONE } else { ZERO };
+    (is_16_or_32, is_16_or_48)
 }