+++ processor/src/decorators/tests.rs
@@ -3,11 +3,11 @@ use super::{
     Process,
 };
 use crate::{MemAdviceProvider, StackInputs, Word};
-use test_utils::{crypto::get_smt_remaining_key, rand::seeded_word};
+use test_utils::rand::seeded_word;
 use vm_core::{
     crypto::{
         hash::{Rpo256, RpoDigest},
-        merkle::{EmptySubtreeRoots, MerkleStore, MerkleTree, NodeIndex},
+        merkle::{EmptySubtreeRoots, MerkleStore, MerkleTree, NodeIndex, TieredSmt},
     },
     utils::IntoBytes,
     AdviceInjector, Decorator, ONE, ZERO,
@@ -60,6 +60,9 @@ fn push_merkle_node() {
     assert_eq!(expected_stack, process.stack.trace_state());
 }
 
+// SMTGET TESTS
+// ================================================================================================
+
 #[test]
 fn push_smtget() {
     // setup the test
@@ -71,13 +74,10 @@ fn push_smtget() {
 
     // check leaves on empty trees
     for depth in [16, 32, 48] {
-        // compute the remaining key
-        let remaining = get_smt_remaining_key(key, depth);
-
         // compute node value
         let depth_element = Felt::from(depth);
         let store = MerkleStore::new();
-        let node = Rpo256::merge_in_domain(&[remaining.into(), value.into()], depth_element);
+        let node = Rpo256::merge_in_domain(&[key.into(), value.into()], depth_element);
 
         // expect absent value with constant depth 16
         let expected = [ZERO, ZERO, ZERO, ZERO, ZERO, ZERO, ZERO, ZERO, ZERO, ONE, ONE];
@@ -86,9 +86,6 @@ fn push_smtget() {
 
     // check leaves inserted on all tiers
     for depth in [16, 32, 48] {
-        // compute the remaining key
-        let remaining = get_smt_remaining_key(key, depth);
-
         // set depth flags
         let is_16_or_32 = (depth == 16 || depth == 32).then_some(ONE).unwrap_or(ZERO);
         let is_16_or_48 = (depth == 16 || depth == 48).then_some(ONE).unwrap_or(ZERO);
@@ -97,7 +94,7 @@ fn push_smtget() {
         let index = key[3].as_int() >> 64 - depth;
         let index = NodeIndex::new(depth, index).unwrap();
         let depth_element = Felt::from(depth);
-        let node = Rpo256::merge_in_domain(&[remaining.into(), value.into()], depth_element);
+        let node = Rpo256::merge_in_domain(&[key.into(), value.into()], depth_element);
 
         // set tier node value and expect the value from the injector
         let mut store = MerkleStore::new();
@@ -108,10 +105,10 @@ fn push_smtget() {
             value[2],
             value[1],
             value[0],
-            remaining[3],
-            remaining[2],
-            remaining[1],
-            remaining[0],
+            key[3],
+            key[2],
+            key[1],
+            key[0],
             is_16_or_32,
             is_16_or_48,
         ];
@@ -144,8 +141,81 @@ fn push_smtget() {
     }
 }
 
+// SMTINSERT TESTS
+// ================================================================================================
+
+#[test]
+fn inject_smtinsert() {
+    let mut smt = TieredSmt::default();
+
+    // --- insert into empty tree ---------------------------------------------
+
+    let raw_a = 0b_01101001_01101100_00011111_11111111_10010110_10010011_11100000_00000000_u64;
+    let key_a = build_key(raw_a);
+    let val_a = [Felt::new(3), Felt::new(5), Felt::new(7), Felt::new(9)];
+
+    // this is a simple insertion at depth 16, and thus the flags should look as follows:
+    let is_update = ZERO;
+    let is_simple_insert = ONE;
+    let is_16_or_32 = ONE;
+    let is_16_or_48 = ONE;
+    let expected_stack = [is_update, is_simple_insert, is_16_or_32, is_16_or_48];
+    let process = prepare_smt_insert(key_a, val_a, &smt, expected_stack.len(), Vec::new());
+    assert_eq!(build_expected(&expected_stack), process.stack.trace_state());
+
+    // --- update same key with different value -------------------------------
+
+    // insert val_a into the tree so that val_b overwrites it
+    smt.insert(key_a.into(), val_a);
+    let val_b = [ONE, ONE, ZERO, ZERO];
+
+    // this is a simple update, and thus the flags should look as follows:
+    let is_update = ONE;
+    let is_16_or_32 = ONE;
+    let is_16_or_48 = ONE;
+
+    // also, the old value should be present in the advice stack:
+    let expected_stack = [
+        val_a[3],
+        val_a[2],
+        val_a[1],
+        val_a[0],
+        is_update,
+        is_16_or_32,
+        is_16_or_48,
+        ZERO,
+    ];
+    let adv_map = vec![build_adv_map_entry(key_a, val_a, 16)];
+    let process = prepare_smt_insert(key_a, val_b, &smt, expected_stack.len(), adv_map);
+    assert_eq!(build_expected(&expected_stack), process.stack.trace_state());
+}
+
+fn prepare_smt_insert(
+    key: Word,
+    value: Word,
+    smt: &TieredSmt,
+    adv_stack_depth: usize,
+    adv_map: Vec<([u8; 32], Vec<Felt>)>,
+) -> Process<MemAdviceProvider> {
+    let root: Word = smt.root().into();
+    let store = MerkleStore::from(smt);
+
+    let stack_inputs = build_stack_inputs(value, key, root);
+    let advice_inputs = AdviceInputs::default().with_merkle_store(store).with_map(adv_map);
+    let mut process = build_process(stack_inputs, advice_inputs);
+
+    process.execute_op(Operation::Noop).unwrap();
+    process
+        .execute_decorator(&Decorator::Advice(AdviceInjector::SmtInsert))
+        .unwrap();
+
+    move_adv_to_stack(&mut process, adv_stack_depth);
+
+    process
+}
+
 // HELPER FUNCTIONS
-// --------------------------------------------------------------------------------------------
+// ================================================================================================
 
 fn init_leaf(value: u64) -> Word {
     [Felt::new(value), Felt::ZERO, Felt::ZERO, Felt::ZERO]
@@ -160,7 +230,7 @@ fn build_expected(values: &[Felt]) -> [Felt; 16] {
 }
 
 fn assert_case_smtget(
-    depth: u8,
+    _depth: u8,
     key: Word,
     value: Word,
     node: RpoDigest,
@@ -169,19 +239,8 @@ fn assert_case_smtget(
     expected_stack: &[Felt],
 ) {
     // build the process
-    let stack_inputs = StackInputs::try_from_values([
-        root[0].as_int(),
-        root[1].as_int(),
-        root[2].as_int(),
-        root[3].as_int(),
-        key[0].as_int(),
-        key[1].as_int(),
-        key[2].as_int(),
-        key[3].as_int(),
-    ])
-    .unwrap();
-    let remaining = get_smt_remaining_key(key, depth);
-    let mapped = remaining.into_iter().chain(value.into_iter()).collect();
+    let stack_inputs = build_stack_inputs(key, root.into(), Word::default());
+    let mapped = key.into_iter().chain(value.into_iter()).collect();
     let advice_inputs = AdviceInputs::default()
         .with_merkle_store(store)
         .with_map([(node.into_bytes(), mapped)]);
@@ -192,13 +251,60 @@ fn assert_case_smtget(
     // call the injector and clear the stack
     process.execute_op(Operation::Noop).unwrap();
     process.execute_decorator(&Decorator::Advice(AdviceInjector::SmtGet)).unwrap();
-    for _ in 0..8 {
+
+    // replace operand stack contents with the data on the advice stack
+    move_adv_to_stack(&mut process, expected_stack.len());
+
+    assert_eq!(build_expected(expected_stack), process.stack.trace_state());
+}
+
+fn build_process(
+    stack_inputs: StackInputs,
+    adv_inputs: AdviceInputs,
+) -> Process<MemAdviceProvider> {
+    let advice_provider = MemAdviceProvider::from(adv_inputs);
+    Process::new(Kernel::default(), stack_inputs, advice_provider, ExecutionOptions::default())
+}
+
+fn build_stack_inputs(w0: Word, w1: Word, w2: Word) -> StackInputs {
+    StackInputs::try_from_values([
+        w2[0].as_int(),
+        w2[1].as_int(),
+        w2[2].as_int(),
+        w2[3].as_int(),
+        w1[0].as_int(),
+        w1[1].as_int(),
+        w1[2].as_int(),
+        w1[3].as_int(),
+        w0[0].as_int(),
+        w0[1].as_int(),
+        w0[2].as_int(),
+        w0[3].as_int(),
+    ])
+    .unwrap()
+}
+
+fn build_key(prefix: u64) -> Word {
+    [ONE, ONE, ONE, Felt::new(prefix)]
+}
+
+/// Removes all items from the operand stack and pushes the specified number of values from
+/// the advice tack onto it.
+fn move_adv_to_stack(process: &mut Process<MemAdviceProvider>, adv_stack_depth: usize) {
+    let stack_depth = process.stack.depth();
+    for _ in 0..stack_depth {
         process.execute_op(Operation::Drop).unwrap();
     }
 
-    // expect the stack output
-    for _ in 0..expected_stack.len() {
+    for _ in 0..adv_stack_depth {
         process.execute_op(Operation::AdvPop).unwrap();
     }
-    assert_eq!(build_expected(expected_stack), process.stack.trace_state());
+}
+
+fn build_adv_map_entry(key: Word, val: Word, depth: u8) -> ([u8; 32], Vec<Felt>) {
+    let node = Rpo256::merge_in_domain(&[key.into(), val.into()], Felt::from(depth));
+    let mut elements = Vec::new();
+    elements.extend_from_slice(&key);
+    elements.extend_from_slice(&val);
+    (node.into(), elements)
 }