+++ processor/src/range/aux_trace.rs
@@ -1,8 +1,6 @@
-use super::{
-    build_lookup_table_row_values, uninit_vector, BTreeMap, ColMatrix, CycleRangeChecks, Felt,
-    FieldElement, RangeCheckFlag, Vec, NUM_RAND_ROWS,
-};
-use miden_air::trace::range::V_COL_IDX;
+use super::{uninit_vector, BTreeMap, ColMatrix, Felt, FieldElement, Vec, NUM_RAND_ROWS};
+use miden_air::trace::range::{M_COL_IDX, V_COL_IDX};
+use vm_core::StarkField;
 
 // AUXILIARY TRACE BUILDER
 // ================================================================================================
@@ -10,16 +8,11 @@ use miden_air::trace::range::V_COL_IDX;
 /// Describes how to construct the execution trace of columns related to the range checker in the
 /// auxiliary segment of the trace. These are used in multiset checks.
 pub struct AuxTraceBuilder {
-    // Range check lookups performed by all user operations, grouped and sorted by clock cycle. Each
-    // cycle is mapped to a single CycleRangeChecks instance which includes lookups from the stack,
-    // memory, or both.
-    // TODO: once we switch to backfilling memory range checks this approach can change to tracking
-    // vectors of hints and rows like in the Stack and Hasher AuxTraceBuilders, and the
-    // CycleRangeChecks struct can be removed.
-    cycle_range_checks: BTreeMap<u32, CycleRangeChecks>,
-    // A trace-length vector of RangeCheckFlags which indicate how many times the range check value
-    // at that row should be included in the trace.
-    row_flags: Vec<RangeCheckFlag>,
+    /// A list of the unique values for which range checks are performed.
+    lookup_values: Vec<u16>,
+    /// Range check lookups performed by all user operations, grouped and sorted by the clock cycle
+    /// at which they are requested.
+    cycle_lookups: BTreeMap<u32, Vec<u16>>,
     // The index of the first row of Range Checker's trace when the padded rows end and values to
     // be range checked start.
     values_start: usize,
@@ -29,21 +22,15 @@ impl AuxTraceBuilder {
     // CONSTRUCTOR
     // --------------------------------------------------------------------------------------------
     pub fn new(
-        cycle_range_checks: BTreeMap<u32, CycleRangeChecks>,
-        row_flags: Vec<RangeCheckFlag>,
+        lookup_values: Vec<u16>,
+        cycle_lookups: BTreeMap<u32, Vec<u16>>,
         values_start: usize,
     ) -> Self {
         Self {
-            cycle_range_checks,
-            row_flags,
+            lookup_values,
+            cycle_lookups,
             values_start,
         }
-    }
-
-    // ACCESSORS
-    // --------------------------------------------------------------------------------------------
-    pub fn cycle_range_check_values(&self) -> Vec<CycleRangeChecks> {
-        self.cycle_range_checks.values().cloned().collect()
     }
 
     // AUX COLUMN BUILDERS
@@ -52,17 +39,14 @@ impl AuxTraceBuilder {
     /// Builds and returns range checker auxiliary trace columns. Currently this consists of two
     /// columns:
     /// - `b_range`: ensures that the range checks performed by the Range Checker match those
-    ///   requested
-    ///    by the Stack and Memory processors.
-    /// - `q`: a helper column of intermediate values to reduce the degree of the constraints for
-    ///    `b_range`. It contains the product of the lookups performed by the Stack at each row.
+    ///   requested by the Stack and Memory processors.
     pub fn build_aux_columns<E: FieldElement<BaseField = Felt>>(
         &self,
         main_trace: &ColMatrix<Felt>,
         rand_elements: &[E],
     ) -> Vec<Vec<E>> {
-        let (b_range, q) = self.build_aux_col_b_range(main_trace, rand_elements);
-        vec![b_range, q]
+        let b_range = self.build_aux_col_b_range(main_trace, rand_elements);
+        vec![b_range]
     }
 
     /// Builds the execution trace of the range check `b_range` and `q` columns which ensure that the
@@ -70,47 +54,39 @@ impl AuxTraceBuilder {
     fn build_aux_col_b_range<E: FieldElement<BaseField = Felt>>(
         &self,
         main_trace: &ColMatrix<Felt>,
-        alphas: &[E],
-    ) -> (Vec<E>, Vec<E>) {
-        // compute the inverses for range checks performed by operations.
-        let (_, inv_row_values) =
-            build_lookup_table_row_values(&self.cycle_range_check_values(), main_trace, alphas);
+        rand_elements: &[E],
+    ) -> Vec<E> {
+        // run batch inversion on the lookup values
+        let divisors = get_divisors(&self.lookup_values, rand_elements[0]);
 
-        // allocate memory for the running product column and set the initial value to ONE
-        let mut q = unsafe { uninit_vector(main_trace.num_rows()) };
+        // allocate memory for the running sum column and set the initial value to ONE
         let mut b_range = unsafe { uninit_vector(main_trace.num_rows()) };
-        q[0] = E::ONE;
         b_range[0] = E::ONE;
 
-        // keep track of the last updated row in the `b_range` running product column. the `q`
-        // column index is always one row behind, since `q` is filled with intermediate values in
-        // the same row as the operation is executed, whereas `b_range` is filled with result
-        // values that are added to the next row after the operation's execution.
+        // keep track of the last updated row in the `b_range` running sum column. `b_range` is
+        // filled with result values that are added to the next row after the operation's execution.
         let mut b_range_idx = 0_usize;
-        // keep track of the next row to be included from the user op range check values.
-        let mut rc_user_op_idx = 0;
 
         // the first half of the trace only includes values from the operations.
-        for (clk, range_checks) in self.cycle_range_checks.range(0..self.values_start as u32) {
+        for (clk, range_checks) in self.cycle_lookups.range(0..self.values_start as u32) {
             let clk = *clk as usize;
 
             // if we skipped some cycles since the last update was processed, values in the last
-            // updated row should by copied over until the current cycle.
+            // updated row should be copied over until the current cycle.
             if b_range_idx < clk {
                 let last_value = b_range[b_range_idx];
                 b_range[(b_range_idx + 1)..=clk].fill(last_value);
-                q[b_range_idx..clk].fill(E::ONE);
             }
 
-            // move the column pointers to the next row.
+            // move the column pointer to the next row.
             b_range_idx = clk + 1;
 
-            // update the intermediate values in the q column.
-            q[clk] = range_checks.to_stack_value(main_trace, alphas);
-
-            // include the operation lookups in the running product.
-            b_range[b_range_idx] = b_range[clk] * inv_row_values[rc_user_op_idx];
-            rc_user_op_idx += 1;
+            b_range[b_range_idx] = b_range[clk];
+            // include the operation lookups
+            for lookup in range_checks.iter() {
+                let value = divisors.get(lookup).expect("invalid lookup value {}");
+                b_range[b_range_idx] -= *value;
+            }
         }
 
         // if we skipped some cycles since the last update was processed, values in the last
@@ -118,13 +94,13 @@ impl AuxTraceBuilder {
         if b_range_idx < self.values_start {
             let last_value = b_range[b_range_idx];
             b_range[(b_range_idx + 1)..=self.values_start].fill(last_value);
-            q[b_range_idx..self.values_start].fill(E::ONE);
         }
 
-        // after the padded section of the range checker table, include `z` in the running product
-        // at each step and remove lookups from user ops at any step where user ops were executed.
-        for (row_idx, (hint, lookup)) in self
-            .row_flags
+        // after the padded section of the range checker table, include the lookup value specified
+        // by the range checker into the running sum at each step, and remove lookups from user ops
+        // at any step where user ops were executed.
+        for (row_idx, (multiplicity, lookup)) in main_trace
+            .get_column(M_COL_IDX)
             .iter()
             .zip(main_trace.get_column(V_COL_IDX).iter())
             .enumerate()
@@ -133,32 +109,64 @@ impl AuxTraceBuilder {
         {
             b_range_idx = row_idx + 1;
 
-            b_range[b_range_idx] = b_range[row_idx] * hint.to_value(*lookup, alphas);
-
-            if let Some(range_check) = self.cycle_range_checks.get(&(row_idx as u32)) {
-                // update the intermediate values in the q column.
-                q[row_idx] = range_check.to_stack_value(main_trace, alphas);
-
-                // include the operation lookups in the running product.
-                b_range[b_range_idx] *= inv_row_values[rc_user_op_idx];
-                rc_user_op_idx += 1;
+            if multiplicity.as_int() != 0 {
+                // add the value in the range checker: multiplicity / (alpha - lookup)
+                let value = divisors.get(&(lookup.as_int() as u16)).expect("invalid lookup value");
+                b_range[b_range_idx] = b_range[row_idx] + value.mul_base(*multiplicity);
             } else {
-                q[row_idx] = E::ONE;
+                b_range[b_range_idx] = b_range[row_idx];
+            }
+
+            // subtract the range checks requested by operations
+            if let Some(range_checks) = self.cycle_lookups.get(&(row_idx as u32)) {
+                for lookup in range_checks.iter() {
+                    let value = divisors.get(lookup).expect("invalid lookup value");
+                    b_range[b_range_idx] -= *value;
+                }
             }
         }
 
         // at this point, all range checks from user operations and the range checker should be
         // matched - so, the last value must be ONE;
-        assert_eq!(q[b_range_idx - 1], E::ONE);
         assert_eq!(b_range[b_range_idx], E::ONE);
 
-        if (b_range_idx - 1) < b_range.len() - 1 {
-            q[b_range_idx..].fill(E::ONE);
-        }
         if b_range_idx < b_range.len() - 1 {
             b_range[(b_range_idx + 1)..].fill(E::ONE);
         }
 
-        (b_range, q)
+        b_range
     }
+}
+
+/// Runs batch inversion on all range check lookup values and returns a map which maps of each value
+/// to the divisor used for including it in the LogUp lookup. In other words, the map contains
+/// mappings of x to 1/(alpha - x).
+fn get_divisors<E: FieldElement<BaseField = Felt>>(
+    lookup_values: &[u16],
+    alpha: E,
+) -> BTreeMap<u16, E> {
+    // run batch inversion on the lookup values
+    let mut values = unsafe { uninit_vector(lookup_values.len()) };
+    let mut inv_values = unsafe { uninit_vector(lookup_values.len()) };
+    let mut log_values = BTreeMap::new();
+
+    let mut acc = E::ONE;
+    for (i, (value, inv_value)) in values.iter_mut().zip(inv_values.iter_mut()).enumerate() {
+        *inv_value = acc;
+        *value = alpha - E::from(lookup_values[i]);
+        acc *= *value;
+    }
+
+    // invert the accumulated product
+    acc = acc.inv();
+
+    // multiply the accumulated product by the original values to compute the inverses, then
+    // build a map of inverses for the lookup values
+    for i in (0..lookup_values.len()).rev() {
+        inv_values[i] *= acc;
+        acc *= values[i];
+        log_values.insert(lookup_values[i], inv_values[i]);
+    }
+
+    log_values
 }